# docs/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hotel competitor monitor API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/settings": {
   "get": {
    "tags": [
     "tenants"
    ],
    "summary": "Get Settings ",
    "operationId": "get_settings__settings_get",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TenantOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "tenants"
    ],
    "summary": "Update Settings",
    "operationId": "update_settings_settings_patch",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TenantUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TenantOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/notifications/settings": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "Get Settings ",
    "operationId": "get_settings__notifications_settings_get",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationSettingsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/recipients": {
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Add Recipient",
    "operationId": "add_recipient_notifications_recipients_post",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RecipientCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RecipientOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/recipients/{recipient_id}": {
   "delete": {
    "tags": [
     "notifications"
    ],
    "summary": "Remove Recipient",
    "operationId": "remove_recipient_notifications_recipients__recipient_id__delete",
    "parameters": [
     {
      "name": "recipient_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Recipient Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/rules/{kind}": {
   "put": {
    "tags": [
     "notifications"
    ],
    "summary": "Update Rule",
    "operationId": "update_rule_notifications_rules__kind__put",
    "parameters": [
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/NotificationKind"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NotificationRuleUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationRuleOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/test": {
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Send Test",
    "operationId": "send_test_notifications_test_post",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationLogOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/log": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "Notification Log",
    "operationId": "notification_log_notifications_log_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/NotificationLogOut"
         },
         "title": "Response Notification Log Notifications Log Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/subscriptions": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "List Subscriptions",
    "description": "Quản trị tenant thấy mọi đăng ký; người xem chỉ thấy đăng ký của mình.",
    "operationId": "list_subscriptions_notifications_subscriptions_get",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SubscriptionOut"
         },
         "title": "Response List Subscriptions Notifications Subscriptions Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Create Subscription",
    "operationId": "create_subscription_notifications_subscriptions_post",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SubscriptionIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SubscriptionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/subscriptions/{sub_id}": {
   "put": {
    "tags": [
     "notifications"
    ],
    "summary": "Update Subscription",
    "operationId": "update_subscription_notifications_subscriptions__sub_id__put",
    "parameters": [
     {
      "name": "sub_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Sub Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SubscriptionIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SubscriptionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "notifications"
    ],
    "summary": "Delete Subscription",
    "operationId": "delete_subscription_notifications_subscriptions__sub_id__delete",
    "parameters": [
     {
      "name": "sub_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Sub Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/{notification_id}/resolve": {
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Resolve Notification",
    "description": "Đánh dấu \"Đã xử lý\" trên dashboard (người đang đăng nhập).",
    "operationId": "resolve_notification_notifications__notification_id__resolve_post",
    "parameters": [
     {
      "name": "notification_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Notification Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationLogOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/notifications/engagement": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "Engagement",
    "description": "Số tin gửi/lỗi/bỏ qua, lượt nhấn, \"Đã xử lý\" và chi phí theo tuần ISO, theo kênh (đo O4).",
    "operationId": "engagement_notifications_engagement_get",
    "parameters": [
     {
      "name": "weeks",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 52,
       "minimum": 1,
       "default": 8,
       "title": "Weeks"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EngagementOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "NotificationSettingsOut": {
    "properties": {
     "email_configured": {
      "type": "boolean",
      "title": "Email Configured"
     },
     "recipients": {
      "items": {
       "$ref": "#/components/schemas/RecipientOut"
      },
      "type": "array",
      "title": "Recipients"
     },
     "rules": {
      "items": {
       "$ref": "#/components/schemas/NotificationRuleOut"
      },
      "type": "array",
      "title": "Rules"
     },
     "zalo_configured": {
      "type": "boolean",
      "title": "Zalo Configured",
      "default": false
     },
     "channels": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Channels",
      "default": [
       "email",
       "zalo",
       "webhook"
      ]
     }
    },
    "type": "object",
    "required": [
     "email_configured",
     "recipients",
     "rules"
    ],
    "title": "NotificationSettingsOut"
   },
...
   "TenantOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone"
     },
     "scan_times": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Scan Times"
     },
     "horizon_days": {
      "type": "integer",
      "title": "Horizon Days"
     },
     "insight_language": {
      "type": "string",
      "title": "Insight Language"
     },
     "insight_hour": {
      "type": "string",
      "title": "Insight Hour"
     },
     "country_code": {
      "type": "string",
      "title": "Country Code"
     },
     "active": {
      "type": "boolean",
      "title": "Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "source_markets": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Source Markets",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "timezone",
     "scan_times",
     "horizon_days",
     "insight_language",
     "insight_hour",
     "country_code",
     "active",
     "created_at"
    ],
    "title": "TenantOut"
   },
...
  }
 }
}